Extract profile section to component.
